# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Domicilia Core API
  version: 0.1.0
  description: |
    Contrato de la API de domicilia-core. Es la FUENTE DE VERDAD para cualquier cliente (el frontend, el agente de IA):
    los clientes tipados se generan de este archivo.

    Convenciones:
    - Identificadores: UUID como texto.
    - Errores: `application/problem+json` (RFC 9457) con `detail` y `request_id`.
    - Listados que crecen (conversaciones, mensajes, contactos): paginación por CURSOR (`limit` + `cursor`); la respuesta
      trae `next_cursor` (null en la última página). El cursor es opaco.
    - Fechas: RFC 3339 en UTC.
    - Cambios parciales (PATCH): un campo ausente no se toca; `null` deja el campo en blanco.

    ESTADO: este contrato cubre por ahora el módulo de WhatsApp (bandejas, contactos, conversaciones y mensajes) y su
    webhook, el módulo de catálogo (categorías, productos, variantes y modificadores), el de pedidos (carrito y su
    máquina de estados), el de pagos (cobro por ePayco y su webhook) y el de promociones (cupones de descuento). El
    resto de la API (usuarios, organizaciones, planes, roles) se documenta aparte; una prueba
    verifica que lo documentado aquí es exactamente lo que el servidor sirve, y valida las respuestas reales contra
    estos esquemas.
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /v1/organizations/{org_id}/cart:
    parameters:
      - $ref: '#/components/parameters/OrgID'
    get:
      tags: [Carrito]
      operationId: getCart
      summary: Ve el carrito propio en esta organización
      description: Lo crea vacío si todavía no existe. Sin permiso de organización — cualquier cliente autenticado puede pedirle a un negocio activo.
      responses:
        '200':
          description: Carrito
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/organizations/{org_id}/cart/items:
    parameters:
      - $ref: '#/components/parameters/OrgID'
    post:
      tags: [Carrito]
      operationId: addCartItem
      summary: Agrega una línea al carrito
      description: |
        El nombre, el precio y los modificadores se resuelven SIEMPRE contra el catálogo real de la organización —
        nunca contra lo que venga en la petición. Un grupo de modificadores obligatorio (min_select >= 1) que
        aplique al producto y no reciba ninguna opción elegida es un 422, aunque el cliente no lo haya tocado.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddCartItemRequest' }
      responses:
        '201':
          description: Carrito con la línea agregada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/organizations/{org_id}/cart/items/{item_id}:
    parameters:
      - $ref: '#/components/parameters/OrgID'
      - $ref: '#/components/parameters/ItemID'
    patch:
      tags: [Carrito]
      operationId: updateCartItemQuantity
      summary: Cambia la cantidad de una línea
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateQuantityRequest' }
      responses:
        '200':
          description: Carrito con la cantidad actualizada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        default: { $ref: '#/components/responses/Problem' }
    delete:
      tags: [Carrito]
      operationId: removeCartItem
      summary: Quita una línea del carrito
      responses:
        '200':
          description: Carrito sin esa línea
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/organizations/{org_id}/cart/place:
    parameters:
      - $ref: '#/components/parameters/OrgID'
    post:
      tags: [Carrito]
      operationId: placeCart
      summary: Confirma el carrito
      description: |
        Congela nombre y precio de cada línea (nunca vuelven a cambiar, aunque el negocio edite su menú después) y
        pasa el pedido a `placed`. Exige al menos una línea. Queda a la espera de pago — ver
        `POST /organizations/{org_id}/orders/{order_id}/pay`.
      responses:
        '200':
          description: Pedido confirmado (placed)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        default: { $ref: '#/components/responses/Problem' }
...
  /v1/carts/mine:
    get:
      tags: [Carrito]
      operationId: listMyCarts
      summary: Mis carritos abiertos, cruzando organizaciones
      description: |
        Un cliente puede tener un carrito abierto por restaurante a la vez (comprar en uno no reemplaza el
        carrito de otro) — esto es lo que un indicador de carrito visible en cualquier ruta del cliente necesita.
        Nunca incluye un carrito vacío. No pagina: a lo sumo un puñado de carritos abiertos por cliente.
      responses:
        '200':
          description: Carritos abiertos
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/CartSummary' } }
        default: { $ref: '#/components/responses/Problem' }
  /v1/orders/mine:
    get:
      tags: [Mis pedidos]
      operationId: listMyOrders
      summary: Lista mis propios pedidos (los ya confirmados, cruzando organizaciones)
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 30 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: Página de pedidos
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderPage' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/orders/{order_id}:
    parameters:
      - $ref: '#/components/parameters/OrderID'
    get:
      tags: [Mis pedidos]
      operationId: getMyOrder
      summary: Lee uno de mis propios pedidos
      description: El de otro cliente, o el carrito (draft) de cualquiera, responde 404 — no 403, para no revelar que el id existe.
      responses:
        '200':
          description: Pedido
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/orders/{order_id}/cancel:
    parameters:
      - $ref: '#/components/parameters/OrderID'
    post:
      tags: [Mis pedidos]
      operationId: cancelMyOrder
      summary: Cancela uno de mis propios pedidos
      description: Solo si su estado todavía lo permite (antes de `preparing`) — si no, 409.
      responses:
        '200':
          description: Pedido cancelado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/orders/{order_id}/retry-payment:
    parameters:
      - $ref: '#/components/parameters/OrderID'
    post:
      tags: [Mis pedidos]
      operationId: retryOrderPayment
      summary: Reintenta el pago de un pedido con el pago fallido
      description: Vuelve el pedido a `placed`, sobre el MISMO pedido — nunca crea uno nuevo. Después, `POST .../pay` inicia un cobro nuevo.
      responses:
        '200':
          description: Pedido de vuelta en placed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/organizations/{org_id}/orders:
    parameters:
      - $ref: '#/components/parameters/OrgID'
    get:
      tags: [Pedidos]
      operationId: listOrders
      summary: Lista los pedidos de la organización (paginado)
      description: Requiere `org.orders.read`. Nunca incluye carritos (draft) de clientes — solo lo ya confirmado.
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [placed, confirmed, payment_failed, accepted, rejected, preparing, out_for_delivery, delivered, cancelled] }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 30 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: Página de pedidos
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderPage' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/organizations/{org_id}/orders/{order_id}:
    parameters:
      - $ref: '#/components/parameters/OrgID'
      - $ref: '#/components/parameters/OrderID'
    get:
      tags: [Pedidos]
      operationId: getOrder
      summary: Lee un pedido de la organización
      description: Requiere `org.orders.read`.
      responses:
        '200':
          description: Pedido
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/organizations/{org_id}/orders/{order_id}/accept:
    parameters:
      - $ref: '#/components/parameters/OrgID'
      - $ref: '#/components/parameters/OrderID'
    post:
      tags: [Pedidos]
      operationId: acceptOrder
      summary: Acepta un pedido pagado (confirmed → accepted)
      description: Requiere `org.orders.manage`.
      responses:
        '200':
          description: Pedido aceptado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/organizations/{org_id}/orders/{order_id}/reject:
    parameters:
      - $ref: '#/components/parameters/OrgID'
      - $ref: '#/components/parameters/OrderID'
    post:
      tags: [Pedidos]
      operationId: rejectOrder
      summary: Rechaza un pedido pagado (confirmed → rejected)
      description: Requiere `org.orders.manage`.
      responses:
        '200':
          description: Pedido rechazado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/organizations/{org_id}/orders/{order_id}/start-preparing:
    parameters:
      - $ref: '#/components/parameters/OrgID'
      - $ref: '#/components/parameters/OrderID'
    post:
      tags: [Pedidos]
      operationId: startPreparingOrder
      summary: Marca un pedido aceptado en preparación (accepted → preparing)
      description: Requiere `org.orders.manage`.
      responses:
        '200':
          description: Pedido en preparación
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/organizations/{org_id}/orders/{order_id}/dispatch:
    parameters:
      - $ref: '#/components/parameters/OrgID'
      - $ref: '#/components/parameters/OrderID'
    post:
      tags: [Pedidos]
      operationId: dispatchOrder
      summary: Despacha un pedido (preparing → out_for_delivery)
      description: Requiere `org.orders.manage`.
      responses:
        '200':
          description: Pedido despachado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        default: { $ref: '#/components/responses/Problem' }
...
  /v1/organizations/{org_id}/orders/{order_id}/cancel:
    parameters:
      - $ref: '#/components/parameters/OrgID'
      - $ref: '#/components/parameters/OrderID'
    post:
      tags: [Pedidos]
      operationId: cancelOrder
      summary: Cancela un pedido desde el negocio
      description: Requiere `org.orders.manage`. Solo antes de `preparing` — si no, 409.
      responses:
        '200':
          description: Pedido cancelado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        default: { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT de auth-domicilia (GoTrue).
  parameters:
    OrgID:
      name: org_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    ItemID:
      name: item_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    OrderID:
      name: order_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Problem:
      description: |
        Error. Códigos habituales: 401 sin credenciales; 402 el plan no incluye la función; 403 sin permiso o organización
        suspendida; 404 no existe (o archivada); 409 choca con el estado actual; 422 entrada inválida; 502 falló Meta;
        503 función sin configurar.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, example: 'about:blank' }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        request_id: { type: string }
...
    CartSummary:
      type: object
      description: Un carrito abierto del cliente en una organización — nunca uno vacío.
      required: [order_id, organization_id, organization_name, organization_slug, item_count, total_cents, updated_at]
      properties:
        order_id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        organization_name: { type: string }
        organization_slug: { type: string }
        item_count: { type: integer }
        total_cents: { type: integer }
        updated_at: { type: string, format: date-time }
...
    OrderModifierSnapshot:
      type: object
      description: Una opción de modificador elegida, con su precio EN EL MOMENTO del pedido — no cambia si el catálogo cambia después.
      required: [modifier_group_id, modifier_group_name, option_id, option_name, price_delta_cents]
      properties:
        modifier_group_id: { type: string, format: uuid }
        modifier_group_name: { type: string }
        option_id: { type: string, format: uuid }
        option_name: { type: string }
        price_delta_cents: { type: integer }
    OrderItem:
      type: object
      description: 'Nombre y precio son un SNAPSHOT: se congelan al confirmar el carrito y no cambian aunque el negocio edite su menú después.'
      required: [id, product_variant_id, name, unit_price_cents, modifiers, unit_total_cents, quantity, line_total_cents, created_at]
      properties:
        id: { type: string, format: uuid }
        product_variant_id: { type: string, format: uuid, nullable: true }
        name: { type: string }
        unit_price_cents: { type: integer, description: 'Precio de la variante sola, sin modificadores.' }
        modifiers: { type: array, items: { $ref: '#/components/schemas/OrderModifierSnapshot' } }
        unit_total_cents: { type: integer, description: unit_price_cents + la suma de los modificadores elegidos. }
        quantity: { type: integer, minimum: 1, maximum: 50 }
        line_total_cents: { type: integer, description: unit_total_cents * quantity. }
        created_at: { type: string, format: date-time }
    Order:
      type: object
      description: |
        Mientras status es `draft`, ES el carrito. Ver docs/ecommerce.md §2.1 para la máquina de estados completa.
      required: [id, organization_id, customer_id, status, items, subtotal_cents, discount_cents, delivery_fee_cents, total_cents, promotion_id, placed_at, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        customer_id: { type: string, format: uuid }
        status:
          type: string
          enum: [draft, placed, confirmed, payment_failed, accepted, rejected, preparing, out_for_delivery, delivered, cancelled]
        items: { type: array, items: { $ref: '#/components/schemas/OrderItem' } }
        subtotal_cents: { type: integer }
        discount_cents: { type: integer, description: El descuento del cupón aplicado (promotion_id) — 0 si no hay ninguno. }
        delivery_fee_cents: { type: integer, description: Siempre 0 hasta que exista el cálculo de envío. }
        total_cents: { type: integer }
        promotion_id: { type: string, format: uuid, nullable: true, description: 'El cupón aplicado al carrito, si hay uno. Se borra solo en cuanto el carrito vuelve a editarse.' }
        placed_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    OrderPage:
      type: object
      required: [items, total, limit, offset]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Order' } }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
...
    AddCartItemRequest:
      type: object
      required: [product_id, variant_id, quantity]
      properties:
        product_id: { type: string, format: uuid }
        variant_id: { type: string, format: uuid }
        quantity: { type: integer, minimum: 1, maximum: 50, default: 1 }
        option_ids:
          type: array
          maxItems: 50
          items: { type: string, format: uuid }
          description: Opciones de modificador elegidas, de cualquier grupo que aplique al producto.
    UpdateQuantityRequest:
      type: object
      required: [quantity]
      properties:
        quantity: { type: integer, minimum: 1, maximum: 50 }
...